Exception testing and final wrapper methods

+++ src/main/java/bananas/ShoppingListDAO.java
@@ -4,6 +4,7 @@ package bananas;
  * Created by Bryan on 6/7/2014.
  **/
 import java.sql.Connection;
+import java.sql.DatabaseMetaData;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -15,65 +16,94 @@ public class ShoppingListDAO implements BananasDAO {
 	private ResultSet rs;
 	private Statement st;
 	private PreparedStatement ps;
+	private DAOUtils daoUtil;
+
+	public ShoppingListDAO(){
+		if(daoUtil == null){
+			daoUtil = new DAOUtils();
+		}
+	}
+
+	public ShoppingListDAO(DAOUtils daoUtil){
+		this.daoUtil = daoUtil;
+	}
+	
+	/*Helper method to check if table exists*/
+	public boolean tableExists(String tableName){
+		ResultSet returned;
+		
+		boolean isTable = false;
+		try {
+			con = daoUtil.getConnection();
+			DatabaseMetaData metadata = con.getMetaData();
+			returned = metadata.getTables(null, null, tableName, null);
+			isTable = returned.next();
+		}catch(SQLException e){
+			return isTable;
+		}
+		return isTable;
+	}
 	/*
 	 *Method to create the table that holds the names of all current shopping lists. It is called from saveListOfItems
 	 *and in turn, calls the method to addToShoppingListNames
 	 * */
-	private void createTableOfShoppingListNames(String listName)throws DAOException{
-		con = DAOUtils.getConnection(con);
+	private void createTableOfShoppingListNames(String listName){ 
+
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("CREATE TABLE IF NOT EXISTS ShoppingListNames (ListName VARCHAR(25) NOT NULL, PRIMARY KEY (ListName))");
 			ps.execute();
-			
+
 		} catch (SQLException e) {
 			throw new DAOException("There was a problem adding the shopping list to the names table.", e);
 		}finally{
-			DAOUtils.closePrepared(ps);
+			daoUtil.closePrepared(ps);
 		}
 		addToShoppingListNames(listName);
 	}
 	/*
 	 * Method to add shopping list name to ShoppingListNames table
 	 * */
-	private void addToShoppingListNames(String listName) throws DAOException{
-		con = DAOUtils.getConnection(con);
+	private void addToShoppingListNames(String listName){
+
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("INSERT INTO ShoppingListNames (ListName) VALUES(?) ON DUPLICATE KEY UPDATE ListName = ListName");
 			ps.setString(1, listName);
 			ps.executeUpdate();
 		} catch (SQLException e) {
 			throw new DAOException("There was a problem adding the Shopping list to the table of Shopping List names.", e);
 		}
-		
+
 	}
 
 	@Override 
 	/*Creates a table for the shopping list and ads the lists name to a table 
 	 *referencing all current lists*/
-	public void createShoppingListTable(String listName) throws DAOException{
-		con = DAOUtils.getConnection(con);
+	public void createShoppingListTable(String listName){
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("CREATE TABLE IF NOT EXISTS " + listName
 					+ "(Name VARCHAR(25) NOT NULL, Quantity DOUBLE, Price DOUBLE, "
 					+ "Unit VARCHAR(25), Category VARCHAR(50),"
 					+ "PRIMARY KEY (Name))");
 			ps.execute();	
-				
+
 		} catch (SQLException e) {
 			throw new DAOException("There was a problem creating the shopping list table.", e);
 		}finally{
-			DAOUtils.closePrepared(ps);
+			//daoUtil.closePrepared(ps);
+			//daoUtil.closeConnection(con);
 		}
 		createTableOfShoppingListNames(listName);
 	}
-	
-	@Override
+
+	@Override 
 	/*
 	 * Calls createShoppingListTable, then iterates over each list item and inserts a row for each item in the 
 	 * array list. If the list item being inserted already exists, it will be replaced
 	 * */
-	public void saveListOfItems(ListOfItems listToSave) throws DAOException{
-		con = DAOUtils.getConnection(con);
+	public void saveListOfItems(ListOfItems listToSave){
 		createShoppingListTable(listToSave.getName());
 		for(ListItem li : listToSave.getList()){
 			String name = li.getName();
@@ -81,8 +111,9 @@ public class ShoppingListDAO implements BananasDAO {
 			String cat = li.getCategory();
 			Double price = li.getPrice();
 			Double qty = li.getQuantity();
-			
+
 			try {
+				con = daoUtil.getConnection();
 				ps = con.prepareStatement("INSERT INTO " + listToSave.getName() + " "
 						+ "(Name, Quantity, Price, Unit, Category) VALUES(?, ? , ? , ? , ?) ON DUPLICATE KEY UPDATE QUANTITY = QUANTITY");
 				ps.setString(1, name);
@@ -91,41 +122,44 @@ public class ShoppingListDAO implements BananasDAO {
 				ps.setString(4, unit);
 				ps.setString(5, cat);
 				ps.executeUpdate();
-				
+
 			} catch (SQLException e) {
 				throw new DAOException("There was a problem saving the shopping list.", e);
 			}finally{
-				DAOUtils.closePrepared(ps);
+				daoUtil.closePrepared(ps);
 				//DAOUtils.closeConn(con);
 				//NOTE: this is causing errors, need connection pool
 			}
 		}//end for-each
 	}//end saveListOfItems
-	
+
 
 	@Override
 	/*
 	 * Returns a Shopping list from the database with the given name
 	 * */
-	
-	public ListOfItems getListOfItems(String listName) throws DAOException{	
-		con = DAOUtils.getConnection(con);
+
+	public ListOfItems getListOfItems(String listName){	
 		ListOfItems result = new ShoppingList(listName);
 		ListItem item;
 		FoodItem foodItem;
+		boolean notTable = tableExists(listName);
+		if(!notTable){
+			throw new DAOException("This table does not exist");
+		}
 		try {
+			con = daoUtil.getConnection();
 			String query = "SELECT * FROM " + listName;
 			st = con.createStatement();
 			rs = st.executeQuery(query);
-			
 		while(rs.next()){
 				String name = rs.getString(1);
-				
+
 				Double qty = rs.getDouble(2);
 				Double price = rs.getDouble(3);
 				String unit = rs.getString(4);
 				String cat = rs.getString(5);
-				
+
 				foodItem = new FoodItem(name);
 				foodItem.setCategory(cat);
 				item = new ListItem.CreateListItem(foodItem).quantity(qty).unit(unit).price(price).create();
@@ -134,12 +168,12 @@ public class ShoppingListDAO implements BananasDAO {
 				result.addItem(item);
 
 		}//end while
-			
+
 		} catch (SQLException e) {
 			throw new DAOException("There was an error getting the list from the database.", e);
 		}finally{
-			DAOUtils.closeResultSet(rs);
-			DAOUtils.closeStatement(st);
+			daoUtil.closeResultSet(rs);
+			daoUtil.closeStatement(st);
 		}
 		return result;
 	}
@@ -147,9 +181,13 @@ public class ShoppingListDAO implements BananasDAO {
 	 * Adds a list item to a shopping list with the specified name
 	 */
 	@Override
-	public void addItemToList(ListItem itemToAdd, String listName) throws DAOException{
-		con = DAOUtils.getConnection(con);
+	public void addItemToList(ListItem itemToAdd, String listName){
+		boolean notTable = tableExists(listName);
+		if(!notTable){
+			throw new DAOException("This table does not exist");
+		}
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("INSERT INTO " + listName + " (Name, Quantity, Price, Unit, Category) VALUES(?, ? , ? , ? , ?)");
 			ps.setString(1, itemToAdd.getName() );
 			ps.setDouble(2, itemToAdd.getQuantity() );
@@ -162,7 +200,7 @@ public class ShoppingListDAO implements BananasDAO {
 			throw new DAOException(e);		
 		}finally{
 			//DAOUtils.closeConn(con);
-			DAOUtils.closePrepared(ps);
+			daoUtil.closePrepared(ps);
 		}
 
 	}
@@ -170,9 +208,13 @@ public class ShoppingListDAO implements BananasDAO {
 	 * Deletes the specified shopping list and deletes its name from the table of shopping list names
 	 */
 	@Override
-	public void deleteList(String listName) throws DAOException{
-		con = DBConnector.getConnection(con);
+	public void deleteList(String listName){
+		boolean notTable = tableExists(listName);
+		if(!notTable){
+			throw new DAOException("This table does not exist");
+		}
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("DROP TABLE " + listName);
 			ps.execute();
 			ps = con.prepareStatement("DELETE FROM ShoppingListNames WHERE ListName = ?");
@@ -182,18 +224,18 @@ public class ShoppingListDAO implements BananasDAO {
 			throw new DAOException("There was a problem deleting the shopping list.", e);
 		}finally{
 			//DAOUtils.closeConn(con);
-			DAOUtils.closePrepared(ps);
+			daoUtil.closePrepared(ps);
 		}
 		}
 	/*
 	 * queries the table of shopping list names and returns an ArrayList containing all current shopping lists
 	 * */
 	@Override
-	public ArrayList<ListOfItems> getAllShoppingLists() throws DAOException{
-		con = DAOUtils.getConnection(con);
+	public ArrayList<ListOfItems> getAllShoppingLists(){
 		ArrayList<ListOfItems> allLists = new ArrayList<ListOfItems>();
 		ResultSet rs2= null;
 		try {
+			con = daoUtil.getConnection();
 			ps = con.prepareStatement("SELECT * FROM ShoppingListNames");
 			rs2 = ps.executeQuery();
 			while(rs2.next()){
@@ -204,18 +246,22 @@ public class ShoppingListDAO implements BananasDAO {
 		} catch (SQLException e) {
 			throw new DAOException("There was a problem retrieving the shopping lists from the database", e);
 		}finally{
-			DAOUtils.closePrepared(ps);
-			DAOUtils.closeResultSet(rs2);
+			daoUtil.closePrepared(ps);
+			daoUtil.closeResultSet(rs2);
 		}
 		return allLists;
 	}
-	
+
 	@Override
-	public void updateList(ListOfItems listToUpdate) throws DAOException{
-		con = DAOUtils.getConnection(con);
-		for(ListItem item : listToUpdate.getList()){
+	public void updateList(ListOfItems listToUpdate){		
+		boolean notTable = tableExists(listToUpdate.getName());
+		if(!notTable){
+			throw new DAOException("This table does not exist");
+		}
 			try {
+				con = daoUtil.getConnection();
 				//iterate over array list and update each 
+				for(ListItem item : listToUpdate.getList()){
 				ps = con.prepareStatement("INSERT INTO " + listToUpdate.getName() +  "(Name, Quantity, Price, Unit, Category) VALUES(?, ? , ? , ? , ?) ON DUPLICATE KEY UPDATE Quantity = ?, Price = ?, Unit = ?, Category = ?");
 				ps.setString(1, item.getName() );
 				ps.setDouble(2, item.getQuantity() );
@@ -227,12 +273,9 @@ public class ShoppingListDAO implements BananasDAO {
 				ps.setString(8, item.getUnit());
 				ps.setString(9, item.getCategory());
 				ps.executeUpdate();
+				}//end for-each
 			} catch (SQLException e) {
-				e.printStackTrace();
 				throw new DAOException("There was a problem updating the shopping list.", e);
 			}
-		}
-		
 	}
-
 }
